bug fix 4 dataset
